+++ lib/theme/createRcaTheme.d.ts
@@ -1,0 +1,4 @@
+import RcaTheme from '.';
+
+export default function createRcaTheme(
+  customTheme: RcaTheme, defaultTheme?: RcaTheme): RcaTheme;